Compute the matrix of p-value.

See here for source: http://www.sthda.com/english/wiki/visualize-correlation-matrix-using-correlogram.

correlation_test(x, confidence_level = 0.95, ...)

Arguments

x

An object of class "DataFrame" or "matrix".

confidence_level

confidence level for the returned confidence interval. Currently only used for the Pearson product moment correlation coefficient if there are at least 4 complete pairs of observations.

...

further arguments to be passed to or from cor.test.

Value

A list containing three matrices; p_value, lower_bound, and upper bound.

See also

Other utils: process_coefficients, process_variable_importances, reduce_cores, remove_variables