Plot gaussian predictions.

Plot gaussian predictions.

plot_predictions_gaussian(y_true, y_pred)

Arguments

y_true

Ground truth (correct) target values.

y_pred

Estimated target values.

Value

A ggplot object. This plot may be rendered by outputting it to the command line or modified using ggplot semantics.

See also

Other plot: plot_coefficients_processed, plot_metrics_binomial_area_under_curve, plot_metrics_gaussian_cor_score, plot_metrics_gaussian_mean_squared_error, plot_metrics_gaussian_r2_score, plot_predictions_binomial, plot_variable_importances_processed