Replicate metrics.

Replicate metrics.

replicate_metrics(fit_model, predict_model, resample, preprocess, measure, X, y,
  train_size = train_size, categorical_variables = NULL,
  n_divisions = 1000, n_iterations = 100, progress_bar = TRUE,
  n_core = 1, foldid = NULL, ...)

Arguments

fit_model

A function; the function for fitting a model to the data.

predict_model

A function; the function for generating predictions from a fitted model.

resample

A function; the function for resampling the data. Defaults to NULL.

preprocess

A function; the function for preprocessing the data. Defaults to NULL.

measure

A function; the function for measuring the results. Defaults to NULL.

X

A matrix; the independent variables.

y

A vector; the dependent variable.

train_size

A numeric vector of length one; specifies what proportion of the data should be used for the training data set. Defaults to 0.667.

categorical_variables

A logical vector; each value TRUE indicates that column in the data.frame is a categorical variable. Defaults to NULL.

n_divisions

An integer vector of length one; specifies the number of times the data should be divided when replicating the error metrics. Defaults to 1000L.

n_iterations

An integer vector of length one; during each division, specifies the number of times the predictions should be replicated. Defaults to 10L.

progress_bar

A logical vector of length one; specifies whether to display a progress bar during calculations. Defaults to TRUE.

n_core

An integer vector of length one; specifies the number of cores to use for this analysis. Currenly only works on Mac OSx and Unix/Linux systems. Defaults to 1L.

foldid

A vector with length equal to length(y) which identifies cases belonging to the same fold.

...

The arguments to be passed to the algorithm specified.

Value

A list of matrixes, the replicated metrics.

See also

Other replicate: replicate_coefficients, replicate_predictions, replicate_variable_importances