Replicate variable importances.

Replicate variable importances.

replicate_variable_importances(fit_model, extract_variable_importances,
  preprocess, X, y, categorical_variables = NULL, n_samples = 1000,
  progress_bar = TRUE, n_core = 1, ...)

Arguments

fit_model

A function; the function for fitting a model to the data.

extract_variable_importances

A function; the function for extracting variable importances from a model.

preprocess

A function; the function for preprocessing the data. Defaults to NULL.

X

A matrix; the independent variables.

y

A vector; the dependent variable.

categorical_variables

A logical vector; each value TRUE indicates that column in the data.frame is a categorical variable. Defaults to NULL.

n_samples

An integer vector of length one; specifies the number of times the coefficients and predictions should be replicated. Defaults to 1000L.

progress_bar

A logical vector of length one; specifies whether to display a progress bar during calculations. Defaults to TRUE.

n_core

An integer vector of length one; specifies the number of cores to use for this analysis. Currenly only works on Mac OSx and Unix/Linux systems. Defaults to 1L.

...

The arguments to be passed to the algorithm specified.

Value

A data.frame, the replicated variable importance scores.

See also

Other replicate: replicate_coefficients, replicate_metrics, replicate_predictions