Set extract variable importances function.

Sets the function responsible for extracting variable importances from a model.

set_extract_variable_importances(algorithm, family)

Arguments

algorithm

A character vector of length one; the algorithm to run on the data. Choices are one of c("glmnet", "random_forest", "support_vector_machine").

family

A character vector of length one; the type of regression to run on the data. Choices are one of c("gaussian", "binomial"). Defaults to "gaussian".

Value

A function; the function for extracting variable importances from a model.

See also

Other setters: set_categorical_variables, set_coefficients_boolean, set_column_names, set_cores, set_dependent_variable, set_extract_coefficients, set_fit_model, set_independent_variables, set_looper_, set_looper, set_measure, set_metrics_boolean, set_parallel, set_plot_metrics, set_plot_predictions, set_predict_model, set_predictions_boolean, set_preprocess, set_random_state, set_resample, set_variable_importances_boolean